requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
EasyProcess==1.1
email_validator==2.1.1
entrypoint2==1.1
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.13.4
frozenlist==1.4.1
gdown==5.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
orjson==3.10.1
patool==2.2.0
pycares==4.4.0
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyparsing==3.0.9
pyrfc6266==1.0.2
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pyunpack==0.3
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
text-unidecode==1.3
tqdm==4.66.2
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
